Added missing createIntegrator() arguments for filters
currentTime_ was not passed to createIntegrator(), resulting in
compilation failure when including the filters.
This commit also includes the necessary changes to the CMake files to
reinclude the filters in compilation, as necessary to compile the
accompanying unit tests.
Since this commit does not fix the
inherent issues with the filtering implementation, it might be useful to
turn filter compilation off again when including it in the actual main
branch.